댓글 삭제 구현

+++ skuhub/src/main/java/org/skuhub/skuhub/api/taxi/dto/request/TaxiCommentDeleteRequest.java
@@ -8,6 +8,6 @@ import java.time.OffsetDateTime;
 @Getter
 @Setter
 public class TaxiCommentDeleteRequest {
-    private int postId;
-    private int commentId;
+    private Long postId;
+    private Long commentId;
 }
